requirements.txt
flask
Flask-SQLAlchemy
Flask-JWT
Flask-Restful
flask-login
openai